requirements.txt
--extra-index-url https://pypi.nvidia.com
cuml-cu12
numpy<2.2
pandas
scikit-learn
torch
torchmetrics
transformers
datasets
matplotlib
tqdm
